Factor out json-based http requests into a common.mjs method
